refactor: create covsirphy.loading package
